+++ examples/resources/coralogix_archive_metrics/resource.tf
@@ -15,6 +15,6 @@ provider "coralogix" {
 resource "coralogix_archive_metrics" "example" {
   s3 = {
     region = "eu-north-1"
-    bucket = "coralogix-c4c-eu2-prometheus-data"
+    bucket = "<your-bucket-name>"
   }
 }