+++ internal/provider/resource_coralogix_recording_rules_groups_set_test.go
@@ -102,6 +102,64 @@ func TestAccCoralogixRecordingRulesGroupsSetFromYamlWithName(t *testing.T) {
 	})
 }
 
+func TestAccCoralogixRecordingRulesGroupsSetUpdateName(t *testing.T) {
+	wd, err := os.Getwd()
+	if err != nil {
+		panic(err)
+	}
+	parent := filepath.Dir(filepath.Dir(wd))
+	filePath := parent + "/examples/resources/coralogix_recording_rules_groups_set/rule-group-set.yaml"
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckRecordingRulesGroupDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccCoralogixResourceRecordingRulesGroupsSetFromYamlWithName(filePath),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(recordingRulesGroupsSetResourceName, "id"),
+					resource.TestCheckResourceAttr(recordingRulesGroupsSetResourceName, "name", "Name"),
+					resource.TestCheckTypeSetElemNestedAttrs(recordingRulesGroupsSetResourceName, "groups.*",
+						map[string]string{
+							"name":     "Foo",
+							"interval": "180",
+							"rules.#":  "2",
+						},
+					),
+					resource.TestCheckTypeSetElemNestedAttrs(recordingRulesGroupsSetResourceName, "groups.*",
+						map[string]string{
+							"name":     "Bar",
+							"interval": "60",
+							"rules.#":  "2",
+						},
+					),
+				),
+			},
+			{
+				Config: testAccCoralogixResourceRecordingRulesGroupsSetFromYamlWithNameUpdated(filePath, "NameUpdated"),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(recordingRulesGroupsSetResourceName, "id"),
+					resource.TestCheckResourceAttr(recordingRulesGroupsSetResourceName, "name", "NameUpdated"),
+					resource.TestCheckTypeSetElemNestedAttrs(recordingRulesGroupsSetResourceName, "groups.*",
+						map[string]string{
+							"name":     "Foo",
+							"interval": "180",
+							"rules.#":  "2",
+						},
+					),
+					resource.TestCheckTypeSetElemNestedAttrs(recordingRulesGroupsSetResourceName, "groups.*",
+						map[string]string{
+							"name":     "Bar",
+							"interval": "60",
+							"rules.#":  "2",
+						},
+					),
+				),
+			},
+		},
+	})
+}
+
 func TestAccCoralogixRecordingRulesGroupsExplicit(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
@@ -159,6 +217,15 @@ func testAccCoralogixResourceRecordingRulesGroupsSetFromYamlWithName(filePath st
 					name = "Name"
 				}
 `, filePath)
+}
+
+func testAccCoralogixResourceRecordingRulesGroupsSetFromYamlWithNameUpdated(filePath, name string) string {
+	return fmt.Sprintf(
+		`resource "coralogix_recording_rules_groups_set" "test" {
+					yaml_content = file("%s")
+					name = %q
+				}
+`, filePath, name)
 }
 
 func testAccCoralogixResourceRecordingRulesGroupsSetFromYaml(filePath string) string {